DDP-8188 | create stategies for bulk cohort creation, create factory for patient filters
